improve i3d.js

+++ scripts/i3d.js
@@ -1,5 +1,11 @@
 #!/usr/bin/env node
 
+const path = require('path');
+
+function rootPath(){ return path.join(__dirname, '..', ...arguments); }
+
+process.chdir(rootPath());
+
 const argi = require('argi');
 
 argi.parse({
@@ -7,12 +13,6 @@ argi.parse({
 		type: 'int',
 		alias: 'v',
 		defaultValue: 1
-	},
-	command: {
-		alias: 'c'
-	},
-	argument: {
-		alias: 'a'
 	}
 });
 
@@ -21,22 +21,22 @@ const options = argi.options.named;
 const i3 = require('i3').createClient();
 const log = new (require('log'))({ tag: 'i3d', defaults: { verbosity: options.verbosity, color: true } });
 
-// i3.command('focus left');
-
-// i3.tree(console.log);
-
-// i3.on('workspace', (evt) => {
-//   console.log('workspace event: ', evt);
-// });
+log(1)('options', argi.options);
 
 const i3d = {
+	rootPath,
 	exit: function(){
 		process.kill(process.pid, 'SIGTERM');
+	},
+	error: function(message, err){
+		log.error(message);
+
+		throw new Error(err);
 	},
 	sendCommand: (command, done = log) => {
 		i3.command(command, (err, result) => {
 			if(err){
-				log.error(`Error running command "${command}"`, err);
+				i3d.error(`Error running command "${command}"`, err);
 
 				return done([]);
 			}
@@ -47,7 +47,7 @@ const i3d = {
 	getWorkspaces: (done = log) => {
 		i3.workspaces((err, workspaces) => {
 			if(err){
-				log.error('Cant get workspaces', err);
+				i3d.error('Cant get workspaces', err);
 
 				return done([]);
 			}
@@ -72,16 +72,35 @@ const i3d = {
 			done(focused, workspaces);
 		});
 	},
+	workspace: {
+		rename: (number, name, color, done = log) => {
+			i3d.getFocusedWorkspace((workspace) => {
+				i3d.sendCommand(`rename workspace number ${workspace.num} to "${number}<span foreground='${color}' weight='heavy'>${number} ${name}</span>"`, done);
+			});
+		}
+	},
+	scratchpad: {
+		show: (name, done) => {
+			i3d.sendCommand(`[con_mark="sp*"] move scratchpad, [con_mark="sp_${name}"] scratchpad show`, done);
+		},
+		hide: (done) => {
+			i3d.sendCommand(`[con_mark="sp*"] move scratchpad`, done)
+		},
+		dump: (done) => {
+			i3d.sendCommand(`[con_mark="sp*"] move to workspace "scratchpad dump", [con_mark="sp*"] floating disable, [con_mark="sp*"] unmark, workspace "scratchpad dump"`, done);
+		}
+	},
 	container: {
-		strafe: (direction = 'next', done) => {
+		strafe: (direction = 'right', done) => {
 			i3d.getFocusedWorkspace((workspace, workspaces) => {
 				let index = workspaces.indexOf(workspace);
 
-				if(direction === 'next'){
+				if(direction === 'right'){
 					++index;
 
 					if(index > i3d.workspaceCount - 1) index = 0;
 				}
+
 				else{
 					--index;
 
@@ -93,28 +112,22 @@ const i3d = {
 				i3d.sendCommand(`move container to workspace number ${number}, workspace number ${number}`, done);
 			});
 		},
-		scratchpad: () => {
-			// exec(`i3-msg '[con_mark="sp_${scratchpadName}"] mark "!sp", [con_mark="sp_${scratchpadName}"] unmark'`, function(){
-			// 	console.log('sp move step 1 ', arguments);
-
-			// 	exec(`sleep .1s && i3-msg 'mark "sp_${scratchpadName}", [con_mark="sp*"] move scratchpad, [con_mark="sp_${scratchpadName}"] scratchpad show'`, function(){
-			// 		console.log('sp move step 2 ', arguments);
-
-			// 		exec(`sleep .1s && i3-msg '[con_mark="!sp"] move to workspace "scratchpad dump", [con_mark="!sp"] floating disable, [con_mark="!sp"] unmark'`, function(){
-			// 			console.log('sp move step 3 ', arguments);
-			// 		});
-			// 	});
-			// });
-		}
-	},
-	showScratchpad: () => {
-		//[con_mark="sp*"] move scratchpad, [con_mark="sp_7"] scratchpad show
+		addToScratchpad: (name, done) => {
+			i3d.sendCommand(`[con_mark="sp_${name}"] mark "!sp", [con_mark="sp_${name}"] unmark`, () => {
+				i3d.sendCommand(`mark "sp_${name}", [con_mark="sp*"] move scratchpad, [con_mark="sp_${name}"] scratchpad show`, () => {
+					i3d.sendCommand(`[con_mark="!sp"] move to workspace "scratchpad dump", [con_mark="!sp"] floating disable, [con_mark="!sp"] unmark`, done);
+				});
+			});
+		},
 	}
 };
 
-if(options.command){
-	if(options.command === 'move-prev') i3d.container.strafe('prev', i3d.exit);
-	else if(options.command === 'move-next') i3d.container.strafe('next', i3d.exit);
+if(argi.options.subCommands){
+	const cmd = argi.options.subCommands;
+
+	if(i3d[cmd[0]] && i3d[cmd[0]][cmd[1]]) i3d[cmd[0]][cmd[1]](cmd[2] ? cmd[2] : i3d.exit, cmd[2] ? i3d.exit : null);
+
+	else i3d.error('Unknown command', cmd);
 }
 
 else{
@@ -123,16 +136,14 @@ else{
 
 		log.info(`STDIN: ${data}`);
 
-		if({ stop: 1, close: 1, exit: 1, kill: 1 }[data]){
-			process.kill(process.pid, 'SIGTERM');
-		}
+		if({ stop: 1, close: 1, exit: 1, kill: 1 }[data]) process.kill(process.pid, 'SIGTERM');
 
-		else if(data === 'p'){
-			i3d.container.strafe('prev');
-		}
+		else if(data === 'csr') i3d.container.strafe('right');
 
-		else if(data === 'n'){
-			i3d.container.strafe();
-		}
+		else if(data === 'csl') i3d.container.strafe('left');
+
+		else if(data.startsWith('cs ')) i3d.scratchpad.show(data.replace('cs ', ''));
+
+		else if(data.startsWith('ss ')) i3d.scratchpad.show(data.replace('ss ', ''));
 	});
 }
